+++ Source/DynamicScene/DynamicObject.js
@@ -7,7 +7,8 @@ define([
         './DynamicProperty',
         './DynamicPositionProperty',
         './DynamicVertexPositionsProperty',
-        './CzmlUnitQuaternion'
+        './CzmlUnitQuaternion',
+        './CzmlCartesian3'
     ], function(
         createGuid,
         DeveloperError,
@@ -16,7 +17,8 @@ define([
         DynamicProperty,
         DynamicPositionProperty,
         DynamicVertexPositionsProperty,
-        CzmlUnitQuaternion) {
+        CzmlUnitQuaternion,
+        CzmlCartesian3) {
     "use strict";
 
     /**
@@ -132,6 +134,13 @@ define([
         this.vertexPositions = undefined;
 
         /**
+         * Gets or sets the suggested initial offset for viewing this object
+         * with the camera.  The offset is defined in the east-north-up reference frame.
+         * @type Cartesian3
+         */
+        this.viewFrom = undefined;
+
+        /**
          * Gets or sets the external document.
          *  @type DynamicExternalDocument
          */
@@ -183,6 +192,34 @@ define([
             dynamicObject.position = position = new DynamicPositionProperty();
         }
         position.processCzmlIntervals(positionData);
+        return propertyCreated;
+    };
+
+    /**
+     * Processes a single CZML packet and merges its data into the provided DynamicObject's viewFrom
+     * property. This method is not normally called directly, but is part of the array of CZML processing
+     * functions that is passed into the DynamicObjectCollection constructor.
+     *
+     * @param {DynamicObject} dynamicObject The DynamicObject which will contain the viewFrom data.
+     * @param {Object} packet The CZML packet to process.
+     * @returns {Boolean} true if the property was newly created while processing the packet, false otherwise.
+     *
+     * @see DynamicProperty
+     * @see DynamicObjectCollection
+     * @see CzmlDefaults#updaters
+     */
+    DynamicObject.processCzmlPacketViewFrom = function(dynamicObject, packet) {
+        var viewFromData = packet.viewFrom;
+        if (typeof viewFromData === 'undefined') {
+            return false;
+        }
+
+        var viewFrom = dynamicObject.viewFrom;
+        var propertyCreated = typeof viewFrom === 'undefined';
+        if (propertyCreated) {
+            dynamicObject.viewFrom = viewFrom = new DynamicProperty(CzmlCartesian3);
+        }
+        viewFrom.processCzmlIntervals(viewFromData);
         return propertyCreated;
     };
 
@@ -287,6 +324,7 @@ define([
         targetObject.position = defaultValue(targetObject.position, objectToMerge.position);
         targetObject.orientation = defaultValue(targetObject.orientation, objectToMerge.orientation);
         targetObject.vertexPositions = defaultValue(targetObject.vertexPositions, objectToMerge.vertexPositions);
+        targetObject.viewFrom = defaultValue(targetObject.viewFrom, objectToMerge.viewFrom);
         targetObject._setAvailability(defaultValue(targetObject.availability, objectToMerge.availability));
     };
 
@@ -303,6 +341,7 @@ define([
         dynamicObject.position = undefined;
         dynamicObject.orientation = undefined;
         dynamicObject.vertexPositions = undefined;
+        dynamicObject.viewFrom = undefined;
         dynamicObject._setAvailability(undefined);
     };