+++ Source/Scene/Camera.js
@@ -153,28 +153,23 @@ define([
      *
      * @memberof Camera
      *
-     * @param {Array} arguments If one parameter is passed to this function, it must have three
-     * properties with the names eye, target, and up; otherwise three arguments are expected which
-     * the same as the properties of one object and given in the order given above.
+     * @param {Cartesian3} eye The position of the camera.
+     * @param {Cartesian3} target The position to look at.
+     * @param {Cartesian3} up The up vector.
      *
+     * @exception {DeveloperError} eye is required.
+     * @exception {DeveloperError} target is required.
+     * @exception {DeveloperError} up is required.
      */
-    Camera.prototype.lookAt = function() {
-        var eye, target, up;
-        if (arguments.length === 1) {
-            var param = arguments[0];
-            if (param.eye && param.target && param.up) {
-                eye = param.eye;
-                target = param.target;
-                up = param.up;
-            } else {
-                return;
-            }
-        } else if (arguments.length === 3) {
-            eye = arguments[0];
-            target = arguments[1];
-            up = arguments[2];
-        } else {
-            return;
+    Camera.prototype.lookAt = function(eye, target, up) {
+        if (typeof eye === 'undefined') {
+            throw new DeveloperError('eye is required');
+        }
+        if (typeof target === 'undefined') {
+            throw new DeveloperError('target is required');
+        }
+        if (typeof up === 'undefined') {
+            throw new DeveloperError('up is required');
         }
 
         this.position = Cartesian3.clone(eye, this.position);
@@ -279,7 +274,7 @@ define([
 
         // Not exactly -z direction because that would lock the camera in place with a constrained z axis.
         this.direction = new Cartesian3(0.0, 0.0001, -0.999);
-        this.right = Cartesian3.UNIT_X;
+        Cartesian3.UNIT_X.clone(this.right);
         this.up = this.right.cross(this.direction);
     };
 
@@ -334,6 +329,10 @@ define([
         this.frustum.left = -right;
         this.frustum.top = top;
         this.frustum.bottom = -top;
+
+        //Orient the camera north.
+        Cartesian3.UNIT_X.clone(this.right);
+        this.up = this.right.cross(this.direction);
     };
 
     function updateViewMatrix(camera) {