+++ CHANGES.md
@@ -7,12 +7,19 @@ Beta Releases
 ### b9 - xx/xx/2012
 
 * Breaking changes:
-   * ...
+   * `shallowEquals` has been removed.
+   * Passing `undefined` to any of the set functions on `Billboard` now throws an exception.
+   * Passing `undefined` to any of the set functions on `Polyline` now throws an exception.
+   * `PolygonPipeline.scaleToGeodeticHeight` now takes ellipsoid as the last parameter, instead of the first.  It also now defaults to `Ellipsoid.WGS84` if no parameter is provided.
 * The new Sandcastle live editor and demo gallery replace the Sandbox and Skeleton examples.
 * Improved picking performance and accuracy.
 * `Extent` functions now take optional result parameters.  Also added `getCenter`, `intersectWith`, and `contains` functions.
+* Add new utility class, `DynamicObjectView` for tracking a DynamicObject with the camera across scene modes; also hooked up CesiumViewerWidget to use it.
+* Added `enableTranslate`, `enableZoom`, and `enableRotate` properties to `Camera2DController` to selectively toggle camera behavior.  All values default to `true`.
+* Added `Camera2DController.setPositionCartographic` to simplify moving the camera programmatically when in 2D mode.
 * Added 'CzmlProcessor' to process multiple CZML fragments/documents. CzmlProcessor also knows how to handle 'external' CZML links and update the documents appropriately based on the CZML properties.
 * Added 'DynamicExternalDocument' to support CZML 'external' property. The external property allows linking with documents that are located at an external url.
+
 ### b8 - 09/05/2012
 
 * Breaking changes: